pyproject.toml
[project]
name = "agentic-ai-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
]
